requirements.txt
flask
openai
langchain
faiss-cpu
requests
python-dotenv
langchain-community
beautifulsoup4
tiktoken
langchain-openai